[INTERNAL] ThemeManager: Support 2nd tier fallback if derived theme is unsupported
In case the fallback SAP theme derived from the theme service URL is
already unsupported, we now "fall up" to the defined default theme.

This will happen only if the theme service itself is unreachable,
as otherwise the theme service has a backend side fallback in place.

JIRA: CPOUI5FRAMEWORK-747
Change-Id: I2b95aae17ee9d35b65f193f19b757057eafdf283