[INTERNAL] plugins.ODataV4Selection: Improve scroll performance
When the table is scrolled, the header selector is updated as part of
the table's update process. Since determining the correct state can be
very time consuming, and hence decreases performance, the "everything is
selected" state is removed.

If there is a selection limit, the corresponding icon is never shown for
hierarchies and with data aggregation. The plugin only visualizes the
"nothing is selected" and the "something is selected" states.

This changes has no effect if the limit is disabled. In this case a
2-state checkbox is displayed and the "everything is selected" state
must be visualized.

In additionan, the performance of range selection is improved.

Change-Id: I26be208ecb76e78a2bde1f77cb44862b8a5de8b3
SNOW: DINC0431510