[INTERNAL] DragAndDrop on mobile devices and ColumnResizer fix
 - Mobile devices needs text data with a value set on the dataTransfer
object to allow dragging. To detect this we now use
Device.browser.mobile which also returns the right value even though the
desktop version is requested on mobile devices
SNOW: DINC0455614

 - ColumnResizer was rendering a <div> to show the handle indicator in
the <table> element which is not compliant with HTML5 specification.
This handle indicator is moved to the <table> container.
SNOW: DINC0452803
Fixes: https://github.com/SAP/openui5/issues/4246

Change-Id: I15345d57a94dc993c44ede22d7eacd71e558d658