[FIX] Demo Kit: Fix Highlighting spaces
Problem: When working with text nodes, its "data"
property will return the text formatted as it is
in the Text Editor, with all the whitespaces and
new lines added by authors keyboard. Using
innerText to then apply that value in the newly
created highlighted Node will be forced to ignore
regex symbols, so the browser rendering is simply
adding their HTML equivalent, which is <br /> tag.
It is a solution, to use textContent to enter the
text into the span element, but it is strongly
recommended we stick to the innerText API, so I am
trimming the possibly problematic symbols manually.

SNOW: DINC0426559
Change-Id: Ie819540f927272a9e2829a259a1056fbcd315d25