[FIX] sap.m.Input: prevent displaying password value when switching type
Previously, when a user set a value and then switched the input type from password to text, the value would briefly become visible for a few milliseconds—posing a potential security risk. With the recent change, the input type is now set synchronously with the value to avoid any mismatch in the visual state. Note that if the value is assigned before the input type is changed, the issue may still occur.

SNOW: DINC0466064
Change-Id: I071ac42255eb66cb815470e40e77e587ef4dbd6b