[FIX] sap.uxap.ObjectPageLayout: Improve accessibility behavior
When the landmarkInfo aggregation of the sap.uxap.ObjectPageLayout
control is used and any of the ObjectPageAccessibleLandmarkInfo
role properties are set to None, all ARIA attributes are removed from
the corresponding element. This behavior occurs, because elements
without an assigned role are not intended to convey semantic meaning.

JIRA: BGSOFUIPIRIN-6821
Change-Id: I31b27396959273c20ad66af6153a7a33b9dd414f