[INTERNAL] ListBase: item should not be hidden under the sticky area
 - When an item is focused the ListBase already calculates whether the
item is already under the sticky area and brings it to the viewport if
needed. In this incident while the item is already focused and new
sticky areas are added to the list. In this case again the sticky focus
handler logic is executed to bring the item in the visible area.

SNOW: CS20250009681888

Change-Id: Iaa466ca52f108dd0e75d6d8b437f9c84578ced8a