[FIX] sap.uxap.ObjectPageLayout: Section title in anchorBar corrected
Problem:
Empty section title in the anchorBar in the case when
the app did NOT set title to that section, but added
to it a single visible subSection that has title.

Solution:
As by UX rule the subSection title should be displayed
in this case, added the required handling.

SNOW: DINC0475577
Change-Id: Id93e37a0d04064d946de663876e5090790851e73