[FIX] sap.m.MultiInput:  Tokenizer Opener is now adjusted
Previously the opener of the tokenizer popover was the tokenizer itself
which caused the popover arrow to move to the right when the tokenizer gets expanded.
Now the opener is set to the MultiInput control which fixes the issue as there is no width change of the opener.

SNOW: DINC0474592
Change-Id: I249c25ed5409e2bfee7fc4e46072286782f713fc