[FIX] sap.uxap.ObjectPageLayout: Fix scrollTop of single SubSection
Problem: "sectionChange" event is not fired when user press the
AnchorBar button to scroll to a Section with a single SubSection.

Solution: Correct the scrollTop position of a single SubSection (within
a certain Section), when it is not promoted - it still has the parent's
scrollTop position as its own scrollTop.

SNOW: DINC0475542
Change-Id: I6f41333cdcc08c010db2e93e7ca0499d68f15c8f