[FIX] Empty Filter for navigation properties
Empty means that no target exists at all, so an All-filter with no result is needed
NotEmpty means, taht there is any target, so an Any-filter without conditions is needed
Exclusive-filters (like NE) means that no target with the filter exists, so an All-filter is needed

The "Authors" sample was enhanced to have an Any-Filter for "Genre" and a corresponding column for the table. Here some fixed in the Delegates are done to support this.

Change-Id: Icb8831c0d367d0cc289cee882fd096d7cc956d74
SNOW: CS20250009421370
SNOW: DINC0467289