[FIX] sap.m.Tokenizer: Tokens are properly updated in the n-more list after delete
Previously, when a token was deleted and the n-more list was expanded,
the deleted token was still visible in the n-more list.
The list was only updated after the n-more popover was re-opened.
Now the list is properly updated after tokenDelete event even
if the popover is open the whole time.

SNOW: DINC0460969
Change-Id: I1b84ee5b6dec5e3cf3dd563adaa570cf7a0c1089