[INTERNAL]sap.ui.rta: Improve change creation on command stack
To ensure that all changes are added to the persistence before a command
is executed, the command stack is now responsible for persisting the
changes. This makes it possible for e.g. MDC Controls to use the
"waitForChanges" method to only update the UI after all changes are
applied. This method cannot be used if the changes are not in the
persistence before the command is executed.

This change also improves the undo logic, which now also waits for the
previous command to be finished before executing the next undo,
preventing possible issues with successive undo operations.

JIRA: TEAM42-414
Change-Id: Id2b5a29e52f7420de4b03894ab7fda779bc3dfe8