[FIX] sap.uxap.ObjectPageLayout: Fix visibility of title IconTabBar mode
Problem: When single SubSection has title in IconTabBar mode, the title
is shown, while it should not be according to the UX spec.

Solution: Set internalTitleVisible property of the SubSection in this
case to false. Internal title visibility logic has been removed from
SubSection level with the redesign of OPL, but it seems it is still
needed in this specific case. So it is now back moved on
ObjectPageSectionBase level, so that it works for both ObjectPageSection
and ObjectPageSubSetion.

SNOW: DINC0493140
Change-Id: Iae4b2e4cb8a17d5700e6618457792128d56ed29a