[INTERNAL][FIX] sap.ui.rta: Fix iframe dialog validation
Previously the "areAllFieldsValid" flag was only calculated in fields
other than the URL field. This caused a situation where the validation
state was invalid when touching other fields without entering the URL
first. Entering the URL afterwards would not trigger a re-validation.

Now the validation is properly handled initially and when the url is
touched (re-checking validation states of other fields as well).

SNOW: DINC0458501
Change-Id: I598bcae9fc577a31b1807b3111ecb26f883da9ae