[INTERNAL] mdc.flexibility.Util: resume control invalidation even after failed appliance
There is a mechanism on p13n side that pauses - and later resumes - control invalidation as long as flex changes are being applied. This way "flickering" effects can be circumvented. However, in some cases p13n / flex change appliance fails. This is intended. In this case, the control invalidation needs to be resumed. This was not the cases, leading to missing / duplicate columns in the UI as well as errors in the TableDelegates.

SNOW: DINC0519388
CR-Id: 002075125800002732042025
Change-Id: I1f8e6b5486caf584773cbbbdce47e25d459ab2d5
(cherry picked from commit 7a33a2053a28f944c9d9ac5eb5b664ae6b80dd56)