[FIX] sap.uxap.ObjectPageHeader: fix overflowing behavior
Problem: objectTitle of ObjectPageHeader is not readable when on phone
size and there is a MenuButton as "action".

Solution: MenuButtons does not overflow, as they do not inherit
sap.m.Button control. The whole ActionSheet logic of ObjectPageHeader is
based on checks if the "actions" are Buttons or
ObjectPageHeaderActionButtons. And as MenuButton is not a Button - they
are not enabled for overflow. As a result there is not enough space for
the objectTitle on a small screen and it gets weirdly visualized.

SNOW: DINC0439863
Change-Id: I026a87161c85827546fae85d2ca8e6dce8a00c83