[FIX] sap.m.PlanningCalendar: Disable navigation past max date
Issue: In some cases the "Next" button won't be disabled in Month view
when the maximum month is already reached.

Solution: When checking if the Next button should be enabled, compare
the max date with the last date of the current month. Also check if the
buttons should be enabled when switching views.

SNOW: CS20240008276460
Change-Id: I989a6361dd8ed184b0c34abad28902af14acc050