[INTERNAL][FIX] ExtendControllerPlugin: Disable for sync views
This commit disables the ExtendControllerPlugin for sync views. The
corresponding entry on the context menu is disabled and has a special
text to inform the user that the plugin is not available for sync views.

CR-Id: 002075125900002163152025
SNOW: DINC0500470
Change-Id: I5b71ed636322d5a1feab1fb22a494ea27c1be1aa
(cherry picked from commit 46c3390a8ab2aa89396c9e008b3832f5e8f6fda5)