[FIX] sap.ui.unified.calendar.Month: Improve day re-rendering peformance
Issue:
- All days are re-rendered when a day is selected.
The special types are calculated for for each day during rendering.
The special types depend on a given date. The special dates of the
control are enhanced via the _getSpecialDates method call as a result
for each day. The special dates enhancement could be a costly operation
when there are a lot of special dates configured by the application.

Solution:
- The special dates enhancement is done once before then control
is rendered as it is not dependant on a given date input. As a result
the rendering of the control is noticeable faster when there are a lot
of special dates provided by the application.

Additionally:
- The special dates enhancement methods are removed from the calendar
and date picker controls are they are not used at all.

SNOW: DINC0355517
CR-Id: 002075125900002311812025
Fixes: https://github.com/SAP/openui5/issues/4167
Change-Id: I1772e370050ff813b04591ee6e433dcb4656bddd