[INTERNAL] SelectionPanel: hide descriptions even when selected
A recently introduced feature changed the filtering behavior on the personalization dialog / SelectionPanel. It introduces a second filter for hiding description columns. The changed filter logic was however wrong, which means: users set the following filters:
- showSelected: true
 - hideDescription: false
The SelectionPanel used the following filter
- showSelected: true
- hideDescription: true
This lead to the issue that wrong columns were shown on the SelectionPanel. The reason for this issue was the logical combination of these filters using model.Filter.

Fix: This change fixes the logical combination of the filters and the conditions under which they are applied. If both filters are true, they are not ANDed instead of ORed. Also, if only one is applied, the second filter is neglected.

CR-Id: 002075125800001257242025
SNOW: DINC0496588
Change-Id: I013a4f43bfb7ba8b2a006b98c58f33f69cc4b201
(cherry picked from commit 12e49548f78bf4e96ff9988dcb7573a789e1d00f)