[FIX] sap.uxap.BlockBase: Error with duplicate ID prevented
Problem:
Error for duplicate ID of a form layout object inside a
BlockBase, after destroying and re-creating a component
containing the given BlockBase;

The issue happens because the form Layout object
with the given ID is not destroyed upon destroying
the BlockBase instance. The object was not destroyed
because is was no longer not part of the BlockBase,
as it was replaced by a different layout object by
the BlockBase itself, during its form adjustment.

Solution:
To prevent memory leak and issues with duplicate IDs,
BloackBase now destroys the obsolete form layout upon
replacing it with a new one.

SNOW: DINC0504299
Change-Id: Icf86510a883d3ddfae6c5cd518475808c815df8d
Fixes: https://github.com/SAP/openui5/issues/4278