[FEATURE] Field, FilterField: render Select for fixed value lists
If only a set of fixed values is allowed for a field and no free-text typing, a Select control should be rendered.
If used as FilterField the dropdown should contain an "empty" row to clear the filter. If this row is selected the FilterField must not have any condition, but show the corresponding text.

To reach this a new FieldSelect control is created that inherits from the Select control but disables it's drowdown. Instead the ValueHelp of the parent Field or FilterField is used.

Change-Id: Ib3aaf35066ef7f40ef78cfad9b5ead22f5ad48a6
JIRA: CPOUIFDENVER-2168
JIRA: CPOUIFDENVER-2175