[FIX] sap.uxap.ObjectPageSection: fix header buttons visibility
Problem: expand/collapse button and pin header buttons are partially
hidden in fiori_3 theme, when ObjectPageLayout does not have anchor bar shown.

Solution: buttons are overlapped by sticky header of the Section. Top
margin is adjusted, so that the header of the Section does not hide
expand/collapse and pin buttons.
Additionally, there was a top border of the Section's heading, which is
now removed, as with previous change (OPL redesign) the top border was
moved on a Section's level.

SNOW: DINC0487600
Change-Id: I0bf5bbc368dab851560f62f6490f42cb9d616df4