[INTERNAL] ColumnResizer: Render the handle only when it is shown
 - The Resize Handle is rendered in the <table> element. Although it
works, rendering a <div> element within the <table> is not allowed and
HTML validators complain about that.
 - To make it fully correct we need to refactor the table DOM structure,
probably we need a wrapper for the <table> element. Besides that,
because of sticky column headers and toolbars we need to handle the
stacking context and this requires direct child selectors to be adapted.
Since table has many subclasses we avoid this option.
 - For the simplicity we render the column resizer handle only when the
user hovers the column border. When the user hover out the borders then
the handle will be removed from the DOM.

SNOW: DINC0490759
Fixes: https://github.com/SAP/openui5/issues/4246

Change-Id: Ie8d3760677ae2eb309fb81e06fd9b42507648530