[INTERNAL] p13n.SelectionPanel: set default value for hideDescriptions only for existing redundant columns
A recently introduced feature changed the filtering behavior on the personalization dialog / SelectionPanel. However, this new filtering behavior for the "hide description" filter should only be used when there entities for which this filter can be applied. This was previously not the case. The filter "hide description" was available and *active* although no description columns were existent. This lead to a wrong filter behavior on the SelectionPanel.

Fix: This change changes the default filter value for this hideDescription filter. It is only true, when description columns are available. otherwise it is false (non-existent) - which align with the previous behavior

CR-Id: 002075125800001256272025
SNOW: DINC0499301
Change-Id: I600e8a1d9bc3d66f45b9c3ec02663e81004fc3f3
(cherry picked from commit 602a6c37dba1a2e4c471b10044fac75dbf657b64)