[FIX] We've now introduced a private read only flag to help the developers determine item sync delegation for the upload complete and item deletion actions
This flag is to let consumers of the control to determine if item handling post upload/deletion has to be handled by the application consuming the control or control handles it internally. This flag could be of use to write unified uploadset implmentation code irrespective of version where change were introduced for item delegation.

SNOW: CS20240008990228
Change-Id: If437a0b5f9d720103671083d39920f51ddd11991