[FIX] sap.uxap.ObjectPageSubSection: prevent duplicate ID error
Problem: An error is thrown, as _headerToolbar aggregation is created
twice with same ID (once on an already destroyed ObjectPageSubSection
instance and second time on a new one with same ID recreated).

How the problem happens:
1. ObjectPageSubSection is destroyed by the app.
2. App is calling _triggerVisibleSubSectionsEvents function OPL.
3. As scroll position of the wrapper is not at the right position at
   this moment (top of the selected Section/SubSection) - OPL scrolls to
   the currently selected Section /this will ensure correct lazyLoading
   and correct SubSections determined as "entered the view port"
4. subSectionEnteredViewPort event is fired for already destroyed
   SubSections /this is another problem, need to be handled separately
5. App is working with the "actions" aggregation of the SubSections,
   entered the view port, some of them already destroyed.
6. As they are destroyed, all their aggregations are also destroyed. So
   _getHeaderToolbar function of ObjectPageSubSection now creates new
   OverflowToolbar and sets it to its aggregation.
7. XML fragment is parsed and ObjectPageSubSection with the same ID
   (like the one already destroyed) is created.
8. It creates a new _headerToolbar with the same ID --> error is thrown
   (here original OverflowToolbar was destroyed when the SubSection was
   destroyed, however, it was created again on the already destroyed
   SubSection)

The ElementsRegistry registers the OverflowToolbar created on a
destroyed SubSection. As it is already destroyed, there is no one now to
destroy this obsolete instance of OverflowToolbar. When a SubSection is
created again with the same ID, OverflowToolbar is created also with the
same ID. However, in the ElementsRegistry there is already an
instance with this ID.

Solution: Check if the ObjectPageSubSection is already destroyed, before
creating _headerToolbar aggregation. If destroyed - does not create an
obsolete aggregation. Make sure .getActions still returns just an empty
array --> no error is thrown here too.

SNOW: DINC0493802
Change-Id: I47fe6405be853b490f2e6872d491f4a70d45a750