[FIX] sap.m.Menu: fixed JS error thrown by MenuWrapper
Issue: In some cases when the id of the MenuItem controls contains
characters that make it invalid CSS selector name, a JS error is
thrown by the onclick method of the Menu's internal MenuWrapper.

Solution: Now the id is escaped with CSS.escape() method and there
is no JS error.

SNOW: DINC0518495
Change-Id: Idd71a44bf06471191c7f53a7b8db129aa043aca1