[INTERNAL] sap/ui/core/demokit/tutorial/odatav4: Remove obsolete integration tests
Only step 8 and 11 of the tutorial have integration tests. The tests
of step 11 are a superset of the tests of step 8, so we keep only the
tests of step 11 and remove the test of step 8.
The OPA tests are neither mentioned in the SDK tutorial nor part of the
tutorials dowload file. They are only used internally and executed via
testsuite.odatav4.qunit.js, so the opaTests.qunit.html file is not
needed to start the tests.

PS1: done

Change-Id: If43b4e0d652d4dec94a46860c0337b4c4eb1c7d5
JIRA: CPOUI5ODATAV4-3015