[FIX] sap.m.Menu: fixed context menu re-opening when it is already open
Issue (reproducible only on Mac): When there is a context menu open,
another right click should open again the context menu on other
position, but on Mac it does not happen, the new context menu is
being opened on the second right click (not on first one).

Solution: There is a check added in the begining of the method
openAsContextMenu, and if the popover is still open, it is closed.

SNOW: n/a
Change-Id: Ifd2e957dd7c9d5c1d539de864d2b5b4fc9e6cb57