[INTERNAL] Demo Kit: improve error handling when sample is broken
Background: When sample is broken and user refreshes the page, the
sample hangs in a constant loading mode.

Implementation: Error handling of broken sample is improved. Now, if for
any reason, the sample cannot be loaded (an error is thrown), we delete
the cahce of the service worker and reload the sample. Warning Dialog is
shown, so that user is informed. This works for XMLViews.

If there is a problem in the controller, we need to reload the whole
DemoKit, as the module loader has already cached the "broken"
controller (returned from the service worker's cache before deletion),
so we need to reload the whole DemoKit. Error Dialog is shown in this
case.

JIRA: BGSOFUIPIRIN-6823
Change-Id: I4b86d986a732ce46e451855b4748562db8534813