[FIX] sap.m.Menu: CSS classes are propagated to the submenus
Issue: After the sap.m.Menu refactoring, the propagation of CSS
classes set to the Menu by addStyleClass() or ToggleStyleClass(), or
removing them by removeStyleClass() methods, were not propagated to
the submenus

Solution: Now the propagation works and submenus have same classes as
the main menu.

For test: use snippix #879198 and try to add/remove/toggle CSS
classes to the main menu.

SNOW: n/a
Change-Id: I68f7110841b735680993db9b46f053045836cdbd