[INTERNAL] Lib.js: Deprecate proxy for library namespaces
In v1: A library that is configured as "apiVersion: 1" creates a set of
proxies for a library's (sub-)namespace in order to detect if enums are
stored in such a namespace. For those enums we automatically catch up a
missing DataType.registerEnum call.

In v2: We assume that each library is cleaned up and we can remove the
support for the (somewhat) complex proxy solution. This change is a
basis for testing this assumption.

JIRA: CPOUI5FRAMEWORK-788
Change-Id: I4fa63731e37e92f9983bf237155abd49153377a8