[INTERNAL] Lib.js: Deprecate proxy for library namespaces
In v1: A library that is configured as "apiVersion: 1" creates a set of
proxies for a library's (sub-)namespace in order to detect if enums are
stored in such a namespace. For those enums we automatically catch up a
missing DataType.registerEnum call.

In v2: We assume that each library is cleaned up and we can remove the
support for the (somewhat) complex proxy solution. This change is a
basis for testing this assumption.

Also add fixes for tests from 57addf7eb713cacb85fb93ebca6529eac45fc54a.

JIRA: CPOUI5FRAMEWORK-788
CR-Id: 002075125800001675972025
SNOW: n/a
Change-Id: I4fa63731e37e92f9983bf237155abd49153377a8
(cherry picked from commit edf913773e152d6a4e709fa2ea743f12289e7595)