[INTERNAL] sap.ui.rta/fl: Add translation text type to annotation rename
This commit adds the translation type to the text for the annotation rename.
It can either be passed by the designtime action as part of the change specific data
(as "textType") or it will default to "XFLD" if not provided.

SNOW: DINC0527933
CR-Id: 002075125800001493392025
Change-Id: I1dc53d848d77697ce2efe4400faed1029105940a
(cherry picked from commit 6837ea988b98582baae2e34786be5954bec4d074)