[FIX] sap.f.DynamicPageHeader: fix headerContentLabel upon snap/unsnap
Problem: When header of the DynamicPage is snapped/unsnapped
headerContentLabel of DynamicPageAccessibleLandmarkInfo is not taken
into account.

Solution: Make sure headerContentLabel custom value is applied whenever
the header of the DynamicPage is snapped/unsnapped. Also make sure, if
DynamicPage is rendered initially with snapped header, the collapsed
header aria-label will be applied.

SNOW: DINC0509897
Change-Id: I68ed62da9834a0ff5108cda4080ecbd90265ee46