[FIX] sap.uxap.ObjectPageLayout: Improve accessibility behavior
This change enhances the behavior introduced in commit I31b27396959273c20ad66af6153a7a33b9dd414f
When the landmarkInfo aggregation of the sap.uxap.ObjectPageLayout control is used, and any of the
ObjectPageAccessibleLandmarkInfo role properties - Navigation, Footer, or Header are set to None,
all ARIA attributes are removed from the corresponding element. This is intentional, as elements
without a defined ARIA role are not meant to convey semantic meaning.

SNOW: CS20240008654628
Change-Id: Iebeb5a6c6c71b8889b0679b2aaa235646a50c708