[FIX] sap.uxap.ObjectPageSubsection: No role for promoted subsection
Problem:
When a SubSection is promoted, the properties of the Section's title are propagated to the promoted SubSection's title. In this case, only the SubSection title is displayed, and the Section uses the SubSection's title as its aria-label. However, the promoted SubSection still technically has a title, which causes role="region" to be set (since SubSections with a title receive this role) without a corresponding label, which results in accessibility validation errors. Assigning the same aria-label to both the Section and the SubSection is not an option, as it would cause screen readers to announce the same label twice.

Solution:
When a SubSection is promoted, we remove it`s role="region". In such
cases, accessibility labeling is handled entirely at the Section level, preventing duplicate announcement and resolving validation issues.

SNOW: DINC0508429
Change-Id: I16644213618fffeda66b588937fb122ea165f7c3