[FIX] sap.m.Tokenizer: aria-controls attribute is now set correctly
This commit fixes the issue where the `aria-controls` attribute was not being set with the correct ID on the `sap.m.Tokenizer` control.
The attribute now reflects the correct ID of the sap.m.ResponsivePopover's content, depending on if it is sap.m.Popover or sap.m.Dialog, ensuring better accessibility compliance.

SNOW: DINC0488566
Change-Id: I9b748e140714d621e9fa1309e8ec35b23fc0dcab