[INTERNAL] sap.m.Select: visible second column value and tooltip
When Select control is diplayed in a two column layout now the value of
the Select includes the "text" and the "additionalText" of the selected
item, separated by "-".

When Select is read-only (editable=false)
there is a default tooltip shown if text is truncated. The default
tooltip is the "text" of the item + "additionalText", if any, separated
by "-".

If the user has already set "tooltip" association, it has precedence
over the default tooltip. In this case, the default tooltip is not
provided.

JIRA: BGSOFUIPIRIN-6836
Change-Id: If07c0d0d5b36804c237abcec52fd880ef8514e89