[FIX] p13n.Engine: provide a way to pass event handler context to stateHandlerRegistry
Issue: events that were attached using Engine.getInstance().attachChange could not be detached using detachChange. This was because oftentimes, handler functions used ".bind(this)" which as such creates an entirely that in return cannot be compared to the internal list of event handlers. Hence, the EventProvider in StateHandlerRegistry did not "know" which item to detach.

Fix: This change introduces a "oListener" parameter for the public functions. It can be used to pass the function context. For this it uses the already existing core APIs from EventProvider.

SNOW: CS20250009526482
Change-Id: Ie6766188b6c4f2b2f5cb323cc5a051b9def96dbe