[FIX] sap.ui.fl: Fix bug where select roles dialog could not be opened when no business role description is present
In Cloud Foundry, when no business role description is available, a context in the contexts response, which is used by the ContextVisibility.controller to show available business roles, will no longer have a description property. This change, is not handled by the ContextVisibility.controller coding and is fixed with this change.

Change-Id: Iddb59d01020fedccb153c758401f35f56c38ed18
CR-Id: 002075125800001862822025
SNOW: CS20250009921397