[INTERNAL] sap.m.MessagePopover: Sample update
The sample was enhanced with command that sends the
focus to the MessagePopover if opened, from any control
on the page.
It would allow visually impaired or keyboard  users to
go back to the MessagePopover in case they were
guided to particular field of interest in some
of its list items.

Change-Id: I42c2548f8aba8b0fd71193478c5443a1243e74cc