[INTERNAL] CellSelector: Propagate event if link is clicked
When a link is CTRL + Clicked in a table with the cell selector active, the propagation was stopped. This prevents the Link press event from firing leading to issues, if users are expecting something to happen.
Therefore, we explicitly check whether the control is a sap.m.Link. If so, propagation should not be stopped.

SNOW: DINC0374258
Change-Id: Id0194637cd171d1c73cf4b66d46b81873227eb51