[INTERNAL] sap.m.PlanningCalendar: fixed Demokit sample
Issue: in the sample, there is a Popover that is being open when
an appointment is clicked, and then when a Delete button inside is
pressed, the selected appointment is being deleted and the Popover is
being closed. But because of rerendering, the page is scrolled to the
top.

Solution: now the Popover is being closed before the appointment
deletion and the issue does not happen.

SNOW: DINC0444306
Change-Id: I17bf0e85ccd70150821146c26dc529b756306235