[INTERNAL] sap.ui.documentation: make DataTables work w/o globals
The DataTables 3rdparty modules use UMD-like modules, which, in the
context of UI5, will fallback to globals (called "Browser" in the UMD
boilerplate code).

In UI5 2.x, globals are no longer supported (esp. not for jQuery) and
instead UI5 module APIs must be used to access jQuery.

This change implements a minimal modification that uses probing require
to retrieve jQuery. This implies that jQuery must be loaded upfront.

SNOW: DINC0530250
Change-Id: Iceeec90746c196cb606631187aa6d771a03af4cf