[INTERNAL] FieldSelect: wait for async formatted text on rendering
If rendered text is different from key it might be determined async.
If selected key is changed the rendering is triggered, here onbeforerendering the old text is invalid.
As is SelectedKey Binding for FieldSelect no ValueHelp is used, the descrtiption in Typeahead needs not to be checked in ConditionType. There is no real autocomplete, so Output cannot be different than formatted condition.

Change-Id: Ie6008a01538c8db5ccf299ba7a7c2512fd866465