[INTERNAL] m.Table: Apply sticky on <thead>
Safari does not apply sticky positioning of the <tr> element on its child elements. This seems to be a bug in Safari. This issue causes MessagePopovers to appear at the incorrect place.
To circumvent this issue, sticky and top can be applied to the <thead> element. This will position child elements such as <th> correctly.

SNOW: DINC0497614
Change-Id: Ie223e5e2a82434c569270eaea60c31956859e6b3