[FIX] sap.m.Menu: fixed context menu auto-closing in some cases
Issue: when a context Menu was opened from a control that is inside
of modal Dialog, it auto-closed immediately because the pointer
element that was used for opening the Menu by was placed as a child
of BODY element. Then a login in a Dialog (and its internal Popup)
decided there is a click outside of the Dialog's (and Menu's) area
and closed the Menu.

Solution: now the pointer element is created as close of the DOM
element that is passed as opener of the menu, and the auto-closing
does not happen anymore as the pointer is inside of the Dialog's
DOM (in the reported case).

SNOW: DINC0529975
Change-Id: I717fae3c8ec28bd2129d0662b37e8d468ca5dd90