[FIX] sap.uxap.ObjectPageSubSection: Fix cloning of "actions"
Problem: When ObjectPageSubSection is created through a shareable
template it gets cloned and during the cloning process "actions"
aggregation (now sent to the "content" aggregation of the private
"_headerToolbar" aggregation) is not replicated --> Buttons are
missing.

Solution: Basic clone function does not clone private aggregations, only
public and forwarded. "actions" aggregation of ObjectPageSubSection does
not use aggregation forwarding, as it needs to add some "private"
controls in the _headerToolbar (not created by application) - title and
spacer. That's why now clone function of the ObjectPageSubSection
provide additional cloning of each Control, part of the "actions"
aggregation and add it to the cloned instance (with exception of the
case when "actions" aggregation comes from binding, as it will be created
and each control will be added later by the binding itself).

SNOW: DINC0556524
Change-Id: I4feabfba74af4ca86dc267a275e0ef7b04ffdd6a
CR-Id: 002075125800002207372025
(cherry picked from commit fa7a494525acd1086ce8506ddaaf1916e45b12d4)