[INTERNAL][FIX] CellSelector: Prevent cell selection when starting from input
If the user tries to select all text in an Input or TextArea, it is very likely that they go outside of the cell to ensure that everything in the textbox is selected. This causes cell selection to select the cells.
To avoid this, a check has been introduced, if the mousedown origin is an InputBase, cell selection is prevented.

SNOW: DINC0399216
Change-Id: I857352db2ef9bc79d29f0383d3544b1e7ece3869