[INTERNAL] Add generic ControlInstances test
Counts the instances of OTHER controls which are also created when an
instance of a control is created. Posts the result to path
"/instance-collector" where it will be received, stored, and compared to
previous instance counts (using the parent git commit). See
/test/_instance_collector in sapui5.runtime repo for the implementation.
When this (internal) tool is not available, nothing happens (no
failure).

In case of increase, the instance-collector tool adds a warning comment
in the gerrit change. It is not the goal to break the build because
often it is justified to create other control instances.

Change-Id: I6b9083f39c7b51381b447f725159d3751c278e2c