[FIX] p13n.xConfigAPI: enhanceConfig with conditional destroy
ISSUE: In certain cases, variants where not reset correctly when switching between multiple variants. E.g. Some columns were still visible although they shouldn't. Also it was not possible for app developers to properly use the xConfig customdata for p13n on some of the controls. In general, the p13n state was not correct when working with variants. This was caused by multiple instances of CustomData with the key "xConfig" as CustomData does not validate for unique keys

FIX: This change introduces a mechanism in which xConfigs are first removed before a new one is added. This way, we can ensure that there is only on xConfig CustomData for a specific control

CR-Id: 002075125800002384742025
SNOW: DINC0522484
Change-Id: Ie99e9c33ebd4fff9330045aff60ffa4d857ab869
(cherry picked from commit c30a8c160fa1d6dc77cf4a90a501f0998d169043)