[INTERNAL][FIX] sap.ui.table.*: Invisible table can request contexts if binding is not suspended
With I8179aedfad8cbbfc18ae379919422956a595da82, the table is prevented
from requesting contests from the binding if the table is invisible.
The use-case was to prevent this for an invisible table with a suspended
binding. Preventing is in case of a non-suspended binding can lead to
issues if the (active) binding expects to receive a getContext call.

Change-Id: I783fbe965624604f86f809b5757829b299447f4f
SNOW: DINC0550995