[FIX] We now ensure that a PDF loads irrespective of the contentType if isTrustedSource is set to true
PDFViewer makes use of chrome://settings/content/pdfDocuments / chrome://flags/#pdf-oopif / isTrustedSource / MimeType to determine if PDF’s can be displayed.

If pdf-oopif is enabled / default we trigger HEAD Request to retrieve the content type and validate type as PDF.
Certain application might not support HEAD Request which results in request to fail and PDF is not displayed.

In ES6 supported version of UI5 we trigger fetch() in other version we trigger XMLHttpRequest for the HEAD Request, In scenarios of HEAD Request failures we now check isTrustedSource = true and enable the display of PDF.

SNOW: CS20250009708968
CR-Id: 002075125800002380312025
Change-Id: I3515e08cbd03d37db7db81a57b746f9024d16d4e