[FIX] sap.m.Wizard: aria-labelledby invisible text is set correctly
Previously, in page mode, the call to setNumberInvisibleText was not
reached, causing ariaLabelledBy to be empty for steps beyond the first.
Moved the method call before the page mode check and adjusted it to use
the current step’s index in aStepPath, as getProgress() returned the
highest visited step, leading to incorrect labeling when navigating
back.

SNOW: DINC0557521
Change-Id: I782cd5159f7826c83d78de243bb93d326980469a