[FIX] sap.tnt.SideNavigation: Padding added at the end of the text
Issue: When the text spans the full length of the item, for example,
when it is truncated, there was no padding at the end.

Solution: Added padding at the end of the text. 

Additionally, padding for navigation groups was removed to resolve
the misalignment of arrows between the groups and items
in the Fiori 3 theme.

SNOW: DINC0563989
Change-Id: Iab7b114f74c4f7add2f48635f5cb66907b16d2e8